# Conflicts:
#	analytics/__pycache__/views.cpython-310.pyc
#	api/__pycache__/models.cpython-310.pyc
#	exercises/__pycache__/models.cpython-310.pyc
#	exercises/__pycache__/views.cpython-310.pyc
#	matches/__pycache__/v_api.cpython-310.pyc
#	matches/__pycache__/views.cpython-310.pyc
#	nanofootball/__pycache__/views.cpython-310.pyc
#	players/__pycache__/models.cpython-310.pyc
#	players/__pycache__/v_api.cpython-310.pyc
#	players/__pycache__/views.cpython-310.pyc
#	version/__pycache__/models.cpython-310.pyc

+++ matches/v_api.py
@@ -8,6 +8,7 @@ from events.models import UserEvent, ClubEvent, EventVideoLink
 from matches.models import UserMatch, ClubMatch, UserProtocol, ClubProtocol
 from references.models import PlayerProtocolStatus
 from players.models import UserPlayer, ClubPlayer
+from nanofootball.views import util_check_access
 
 
 LANG_CODE_DEFAULT = "en"
@@ -355,9 +356,11 @@ def POST_edit_match(request, cur_user, cur_team):
     except:
         pass
     c_match = None
-    access_denied = False
     c_team = None
-    if access_denied:
+    if not util_check_access(cur_user, {
+        'perms_user': ["matches.change_usermatch", "matches.add_usermatch"], 
+        'perms_club': ["matches.change_clubmatch", "matches.add_clubmatch"]
+    }):
         return JsonResponse({"err": "Access denied.", "success": False}, status=400)
     post_data = request.POST.get("data", None)
     try:
@@ -368,23 +371,45 @@ def POST_edit_match(request, cur_user, cur_team):
         return JsonResponse({"errors": "Can't parse post data"}, status=400)
     adding_mode = False
     c_datetime = set_value_as_datetime(f"{post_data['date']} {post_data['time']}:00")
-    c_match = UserMatch.objects.filter(event_id=match_id, team_id=cur_team)
-    if not c_match.exists() or c_match[0].event_id == None:
-        adding_mode = True
-        c_team = UserTeam.objects.filter(id=cur_team)
-        if not c_team.exists() or c_team[0].id == None:
-            return JsonResponse({"err": "Team not found.", "success": False}, status=400)
-        try:
-            if c_datetime:
-                new_event = UserEvent(user_id=cur_user, date=c_datetime)
-            else:
-                new_event = UserEvent(user_id=cur_user)
-            new_event.save()
-        except:
-            return JsonResponse({"errors": "Can't create event"}, status=400)
-        c_match = UserMatch(team_id=c_team[0], event_id=new_event)
+    c_match = None
+    if request.user.club_id is not None:
+        c_match = ClubMatch.objects.filter(event_id=match_id, team_id=cur_team)
+        if not c_match.exists() or c_match[0].event_id == None:
+            adding_mode = True
+            c_team = ClubTeam.objects.filter(id=cur_team)
+            if not c_team.exists() or c_team[0].id == None:
+                return JsonResponse({"err": "Team not found.", "success": False}, status=400)
+            try:
+                if c_datetime:
+                    new_event = ClubEvent(user_id=cur_user, date=c_datetime, club_id=request.user.club_id)
+                else:
+                    new_event = ClubEvent(user_id=cur_user, club_id=request.user.club_id)
+                new_event.save()
+            except:
+                return JsonResponse({"errors": "Can't create event"}, status=400)
+            c_match = ClubMatch(team_id=c_team[0], event_id=new_event)
+        else:
+            c_match = c_match[0]
     else:
-        c_match = c_match[0]
+        c_match = UserMatch.objects.filter(event_id=match_id, team_id=cur_team)
+        if not c_match.exists() or c_match[0].event_id == None:
+            adding_mode = True
+            c_team = UserTeam.objects.filter(id=cur_team)
+            if not c_team.exists() or c_team[0].id == None:
+                return JsonResponse({"err": "Team not found.", "success": False}, status=400)
+            try:
+                if c_datetime:
+                    new_event = UserEvent(user_id=cur_user, date=c_datetime)
+                else:
+                    new_event = UserEvent(user_id=cur_user)
+                new_event.save()
+            except:
+                return JsonResponse({"errors": "Can't create event"}, status=400)
+            c_match = UserMatch(team_id=c_team[0], event_id=new_event)
+        else:
+            c_match = c_match[0]
+    if c_match == None:
+        return JsonResponse({"err": "Match not found.", "success": False}, status=400)
     if c_datetime:
         c_match.event_id.date = c_datetime
     c_match.duration = set_value_as_duration(post_data['duration'])
@@ -397,8 +422,6 @@ def POST_edit_match(request, cur_user, cur_team):
     c_match.tournament = post_data['tournament']
     c_match.m_type = set_value_as_int(post_data['m_type'], 0)
     c_match.m_format = post_data['m_format']
-    if c_match == None:
-        return JsonResponse({"err": "Match not found.", "success": False}, status=400)
     try:
         c_match.save()
         res_data = f'Match with id: [{c_match.event_id}] is added / edited successfully.'
@@ -427,10 +450,14 @@ def POST_delete_match(request, cur_user, cur_team):
     except:
         pass
     c_event = None # deleting the event, delete and match
-    access_denied = False
-    if access_denied:
+    if not util_check_access(cur_user, {
+        'perms_user': ["matches.delete_usermatch"], 
+        'perms_club': ["matches.delete_clubmatch"]
+    }):
         return JsonResponse({"err": "Access denied.", "success": False}, status=400)
-    if not access_denied:
+    if request.user.club_id is not None:
+        c_event = ClubEvent.objects.filter(id=match_id, club_id=request.user.club_id)
+    else:
         c_event = UserEvent.objects.filter(id=match_id, user_id=cur_user)
     if c_event == None or not c_event.exists() or c_event[0].id == None:
         return JsonResponse({"errors": "access_error"}, status=400)
@@ -466,12 +493,17 @@ def POST_edit_players_protocol(request, cur_user):
         pass
     c_key = request.POST.get("key", "")
     keys_with_values = ['minute_from', 'minute_to', 'goal', 'penalty', 'p_pass', 'yellow_card', 'red_card', 'estimation', 'like', 'dislike']
+    AnyProtocol = None
+    if request.user.club_id is not None:
+        AnyProtocol = ClubProtocol
+    else:
+        AnyProtocol = UserProtocol
     try:
         update_dict = {
             f'{c_key}': c_value
         }
         if c_key in keys_with_values:
-            f_protocol = UserProtocol.objects.filter(id=protocol_id)
+            f_protocol = AnyProtocol.objects.filter(id=protocol_id)
             if f_protocol.exists() and f_protocol[0].id != None:
                 if f_protocol[0].p_status and 'matches_reset' in f_protocol[0].p_status.tags and f_protocol[0].p_status.tags['matches_reset'] == 1:
                     c_value = None
@@ -504,17 +536,17 @@ def POST_edit_players_protocol(request, cur_user):
                     else:
                         update_dict[t_key] = False
         elif c_key == "is_captain" or c_key == "is_goalkeeper":
-            f_protocol = UserProtocol.objects.filter(id=protocol_id)
+            f_protocol = AnyProtocol.objects.filter(id=protocol_id)
             if f_protocol.exists() and f_protocol[0].id != None:
                 t_val = getattr(f_protocol[0], c_key)
                 update_dict[c_key] = not t_val
         elif c_key == "border_black" or c_key == "border_red":
-            f_protocol = UserProtocol.objects.filter(id=protocol_id)
+            f_protocol = AnyProtocol.objects.filter(id=protocol_id)
             if f_protocol.exists() and f_protocol[0].id != None:
                 t_val = getattr(f_protocol[0], c_key)
                 t_val = 1 if t_val == 0 else 0
                 update_dict[c_key] = t_val
-        UserProtocol.objects.filter(id=protocol_id).update(**update_dict)
+        AnyProtocol.objects.filter(id=protocol_id).update(**update_dict)
     except Exception as e:
         print(e)
         return JsonResponse({"err": "Can't edit match protocol.", "success": False}, status=400)
@@ -567,10 +599,21 @@ def POST_add_delete_players_protocol(request, cur_user, to_add = True):
         if not pl_id:
             continue
         if to_add:
-            f_player = UserPlayer.objects.filter(id=pl_id, user=cur_user, team=team_id)
-            f_match = UserMatch.objects.filter(event_id=match_id)
+            f_player = None
+            f_match = None
+            if request.user.club_id is not None:
+                f_player = ClubPlayer.objects.filter(id=pl_id, team=team_id)
+                f_match = ClubMatch.objects.filter(event_id=match_id)
+            else:
+                f_player = UserPlayer.objects.filter(id=pl_id, user=cur_user, team=team_id)
+                f_match = UserMatch.objects.filter(event_id=match_id)
             if f_player.exists() and f_player[0].id != None and f_match.exists() and f_match[0].event_id != None:
-                protocol, created = UserProtocol.objects.get_or_create(match=f_match[0], player=f_player[0])
+                protocol = None
+                created = False
+                if request.user.club_id is not None:
+                    protocol, created = ClubProtocol.objects.get_or_create(match=f_match[0], player=f_player[0])
+                else:
+                    protocol, created = UserProtocol.objects.get_or_create(match=f_match[0], player=f_player[0])
                 if created:
                     try:
                         protocol.is_opponent = is_opponent
@@ -581,8 +624,12 @@ def POST_add_delete_players_protocol(request, cur_user, to_add = True):
                 else:
                     res_data.append(f"Protocol with id: {protocol.id} (refer on player: {f_player[0].id} and on match: {f_match[0].event_id}) already existed.")
         else:
-            f_protocol = UserProtocol.objects.filter(id=pl_id)
-            if f_protocol.exists() and f_protocol[0].id != None:
+            f_protocol = None
+            if request.user.club_id is not None:
+                f_protocol = ClubProtocol.objects.filter(id=pl_id)
+            else:
+                f_protocol = UserProtocol.objects.filter(id=pl_id)
+            if f_protocol != None and f_protocol.exists() and f_protocol[0].id != None:
                 try:
                     f_protocol[0].delete()
                     res_data.append(f"Protocol was deleted.")
@@ -620,7 +667,11 @@ def POST_edit_players_protocol_order(request, cur_user):
             t_id = int(ids_data[c_ind])
         except:
             pass
-        found_protocol = UserProtocol.objects.get(id=t_id)
+        found_protocol = None
+        if request.user.club_id is not None:
+            found_protocol = ClubProtocol.objects.get(id=t_id)
+        else:
+            found_protocol = UserProtocol.objects.get(id=t_id)
         if found_protocol and found_protocol.id != None:
             found_protocol.order = t_order
             try:
@@ -657,8 +708,12 @@ def POST_edit_match_video_event(request, cur_user, cur_team):
     res_data = "ERROR."
     success_state = False
     status_state = 400
-    f_event = UserEvent.objects.filter(id=event_id)
-    if f_event.exists() and f_event[0].id != None:
+    f_event = None
+    if request.user.club_id is not None:
+        f_event = ClubEvent.objects.filter(id=event_id)
+    else:
+        f_event = UserEvent.objects.filter(id=event_id)
+    if f_event != None and f_event.exists() and f_event[0].id != None:
         f_event_video = f_event[0].video_link
         if f_event_video == None:
             f_event_video = EventVideoLink()
@@ -704,7 +759,11 @@ def POST_edit_match_video_protocol(request, cur_user, cur_team):
     res_data = "ERROR."
     success_state = False
     status_state = 400
-    f_protocol = UserProtocol.objects.filter(id=protocol_id)
+    f_protocol = None
+    if request.user.club_id is not None:
+        f_protocol = ClubProtocol.objects.filter(id=protocol_id)
+    else:
+        f_protocol = UserProtocol.objects.filter(id=protocol_id)
     if f_protocol.exists() and f_protocol[0].id != None:
         f_protocol_video = f_protocol[0].video_link
         if f_protocol_video == None:
@@ -752,7 +811,19 @@ def GET_get_match(request, cur_user, cur_team, return_JsonResponse=True):
     except:
         pass
     res_data = {}
-    match = UserMatch.objects.filter(event_id=match_id, team_id=cur_team)
+    match = None
+    if not util_check_access(cur_user, {
+        'perms_user': ["matches.view_usermatch"], 
+        'perms_club': ["matches.view_clubmatch"]
+    }):
+        if return_JsonResponse:
+            return JsonResponse({"err": "Access denied.", "success": False}, status=400)
+        else:
+            return None
+    if request.user.club_id is not None:
+        match = ClubMatch.objects.filter(event_id=match_id, team_id=cur_team)
+    else:
+        match = UserMatch.objects.filter(event_id=match_id, team_id=cur_team)
     if match.exists() and match[0].event_id != None:
         res_data = match.values()[0]
         res_data['date'] = get_date_str_from_datetime(match[0].event_id.date, LANG_CODE_DEFAULT)
@@ -794,7 +865,11 @@ def GET_get_match_protocol(request, cur_user, cur_team):
     except:
         pass
     res_data = []
-    protocol = UserProtocol.objects.filter(match=match_id)
+    protocol = None
+    if request.user.club_id is not None:
+        protocol = ClubProtocol.objects.filter(match=match_id)
+    else:
+        protocol = UserProtocol.objects.filter(match=match_id)
     if protocol.exists() and protocol[0].id != None:
         for protocol_elem in protocol:
             protocol_dict = model_to_dict(protocol_elem)
@@ -838,7 +913,11 @@ def GET_get_match_video_event(request, cur_user, cur_team, returnJSONResponse=Tr
         pass
     if custom_id:
         event_id = custom_id
-    event = UserEvent.objects.filter(id=event_id)
+    event = None
+    if request.user.club_id is not None:
+        event = ClubEvent.objects.filter(id=event_id)
+    else:
+        event = UserEvent.objects.filter(id=event_id)
     if event.exists() and event[0].id != None:
         res_data = {'links': [], 'notes': []}
         try:
@@ -883,7 +962,11 @@ def GET_get_match_video_protocol(request, cur_user, cur_team, returnJSONResponse
         pass
     if custom_id:
         protocol_id = custom_id
-    protocol = UserProtocol.objects.filter(id=protocol_id)
+    protocol = None
+    if request.user.club_id is not None:
+        protocol = ClubProtocol.objects.filter(id=protocol_id)
+    else:
+        protocol = UserProtocol.objects.filter(id=protocol_id)
     if protocol.exists() and protocol[0].id != None:
         res_data = {'links': [], 'notes': []}
         try:
@@ -899,6 +982,4 @@ def GET_get_match_video_protocol(request, cur_user, cur_team, returnJSONResponse
         return JsonResponse({"errors": "Protocol video not found.", "success": False}, status=400)
     else:
         return None
-
-
